Final changes for Version 2 deployment

+++ plugins/bl/regionalinvestment/updates/builder_table_create_bl_regionalinvestment_business_types.php
@@ -1,5 +1,8 @@
-<?php namespace BL\RegionalInvestment\Updates;
+<?php
 
+namespace BL\RegionalInvestment\Updates;
+
+use Illuminate\Support\Facades\DB;
 use Schema;
 use October\Rain\Database\Updates\Migration;
 
@@ -7,16 +10,23 @@ class BuilderTableCreateBlRegionalinvestmentBusinessTypes extends Migration
 {
     public function up()
     {
-        Schema::create('bl_regionalinvestment_business_types', function($table)
-        {
-            $table->engine = 'InnoDB';
-            $table->bigIncrements('id')->unsigned();
-            $table->string('name');
-        });
+        if (!Schema::hasTable('bl_regionalinvestment_business_types')) {
+            Schema::create('bl_regionalinvestment_business_types', function ($table) {
+                $table->engine = 'InnoDB';
+                $table->bigIncrements('id')->unsigned();
+                $table->string('slug');
+                $table->string('name');
+                $table->integer('sort_order')->nullable()->unsigned();
+            });
+        }
     }
-    
+
     public function down()
     {
-        Schema::dropIfExists('bl_regionalinvestment_business_types');
+        if (Schema::hasTable('bl_regionalinvestment_business_types')) {
+            DB::statement('SET FOREIGN_KEY_CHECKS = 0');
+            Schema::dropIfExists('bl_regionalinvestment_business_types');
+            DB::statement('SET FOREIGN_KEY_CHECKS = 1');
+        }
     }
 }
